requirements.txt
Flask==1.1.1
redis==3.4.1
pybase62==0.4.3